Banner and Category Updates # Updates the banner to be fully responsive # Category cards now change font size with the resp. width # Some final changes pending
