+++ src/components/cards/category_cards.css
@@ -1,61 +1,44 @@
-.category-listing{
-    padding: 10px;
-    display: flex;
-    overflow-x: auto;
-}
 
 .category-card{
-    padding: 8px;
+    --border-radius: 1rem;
     position: relative;
-    min-width: 310px;
-    margin: 10px;
-    border-radius: 22px;
-    border: solid 5px black;
-    cursor: pointer;
+    border: 5px solid black;
+    padding: 1rem;
+    /* margin: 2rem; */
+    width: max(15rem, 13vw);
+    border-radius: var(--border-radius);
 }
 
-.category-card:hover{
-    box-shadow: -8px -8px red;
-    background-color: black;
-    color: white;
-}
-
-.category-card:hover .category-card-number{
-    color: white;
-}
-
-.category-card:hover .category-img > img{
-    border-color: white;  
-}
-
-
-.category-img{
-    width: 300px;
-    height: 350px;
-}
-
-.category-img > img{
+.category-card > img{
     width: 100%;
-    height: 100%;
-    object-fit: cover;
-    border: solid 5px black;
-    border-top-left-radius: 20px;
-    border-top-right-radius: 20px;
+    border-top-right-radius: var(--border-radius);
+    border-top-left-radius: var(--border-radius);
 }
 
-.category-card-number{
+.category-card > h3{
+    /* border: solid 3px red; */
     position: absolute;
-    margin: 0px;
-    left: 20px;
-    top: 10px;
-    font-size: 70px;
+    width: 100%;
+    top: 1rem;
+    margin: 0;
+    left: 2rem;
+    font-size: min(3vw, 5rem);
     color: transparent;
     -webkit-text-stroke: 2px white;
 }
 
-.category-title{
-    margin: 0px;
-    padding: 20px 15px;
+.category-card > h2{
+    /* border: solid 3px hotpink; */
+    bottom: 2rem;
+    left: 2rem;
+    font-size: min(2vw, 3rem);
     text-align: center;
-    font-size: 35px;
+    margin: 1rem 0;
 }
+
+.category-card:hover{
+    background-color: black;
+    color: white;
+}
+
+
