+++ src/components/main_components/banner.css
@@ -1,33 +1,11 @@
-.banner{
-    /* background-image: url(/public/adidas-banner.jpg); */
-    /* background-size:contain; */
-    height: 720px;
-    z-index: 2;
-    margin: 0px 25px;
-    
+section{
+    padding: 0 2.5rem ;
 }
 
 .banner-img{
-    border: solid 5px white;
-    object-fit: cover;
     width: 100%;
-    height: 100%;
+    max-height: 75vh;
+    object-fit: cover;
+    border: solid 5px white;
     box-shadow: 10px 10px black;
 }
-
-/* .header-text{
-    position: relative;
-    left: 3rem;
-    top: 5rem;
-    color: white;
-    width: 25%;
-    
-} */
-/* .header-text h1 {
-    margin-top: 0;
-    font-size: 55px;
-} */
-
-/* .header-text p{
-    font-size: 24px;
-} */
