--- src/components/main_components/banner.js
@@ -1,16 +1,0 @@
-import React from 'react';
-import './banner.css'
-
-function Header() {
-    return (
-        <div className="banner">
-            <img className="banner-img" src="adidas-banner.jpg" alt=""></img>
-            {/* <div className='header-text'>
-                <h1>Shopping And Department Store.</h1>
-                <p>Lorem ipsum dolor sit amet, consectetur adipisicing elit. Consequatur, quis! Nihil placeat nostrum quis vel!consectetur adipisicing elit.</p>
-            </div> */}
-        </div>
-    );
-}
-
-export default Header;
