+++ src/components/main_components/banner.jsx
@@ -1,0 +1,12 @@
+import React from 'react';
+import './banner.css'
+
+function Header() {
+    return (
+        <section>
+            <img className="banner-img" src="adidas-banner.jpg" alt=""></img>
+        </section>
+    );
+}
+
+export default Header;
