+++ src/components/main_components/category_card_listing.jsx
@@ -2,6 +2,8 @@ import React from 'react';
 import CategoryCard from '../cards/category_cards'
 import Listing from '../minor_components/listing'
 import data from '../../data/categories'
+import Heading from '../minor_components/heading';
+import './category_card_listing.css'
 
 function CategoryCardListing() {
     const cardsss = data.map(
@@ -10,7 +12,13 @@ function CategoryCardListing() {
         }
     )
     return (
-        <Listing cards={cardsss} title="Shop Our Top Categories"/>
+        <article>
+            <Heading text={'Shop Our Categories'}/>
+        <section className='category-cards'>
+            {cardsss}
+        </section>
+        </article>
+        // <Listing cards={cardsss} title="Shop Our Top Categories"/>
     );
 }
